feat: custom menu

+++ src/services/mongoose/resto/custom-menu/index.ts
@@ -1,61 +1,24 @@
 import { Request } from 'express';
 import { z } from 'zod';
 
-import CustomMenuCategory, { ICustomMenuCategory } from '../../../models/CustomMenuCategory';
-import { IRestaurant } from '../../../models/Restaurant';
-import CustomMenuCategorySpicyLevel from '../../../models/CustomMenuCategorySpicyLevel';
-import { BadRequest, NotFound } from '../../../errors';
-import CustomMenuComposition, { ICustomMenuComposition } from '../../../models/CustomMenuComposition';
-import convertImageGallery from '../../../utils/convertImageGallery';
+import CustomMenuCategory, { ICustomMenuCategory } from '../../../../models/CustomMenuCategory';
+import { IRestaurant } from '../../../../models/Restaurant';
+import CustomMenuCategorySpicyLevel from '../../../../models/CustomMenuCategorySpicyLevel';
+import { BadRequest, NotFound } from '../../../../errors';
+import CustomMenuComposition, { ICustomMenuComposition } from '../../../../models/CustomMenuComposition';
+import convertImageGallery from '../../../../utils/convertImageGallery';
 
-const customMenuCategoryBodySchema = z.object({
-  name: z.string().nonempty().max(50),
-  isBungkusAble: z.boolean().default(false).optional(),
-  maxSpicy: z.number().nullable().default(null).optional(),
-});
-
-export type CustomMenuCategoryBodyDTO = z.infer<typeof customMenuCategoryBodySchema>;
-export type CustomMenuCategoryResponseDTO = {
-  _id: ICustomMenuCategory['_id'];
-  name: CustomMenuCategoryBodyDTO['name'];
-  isBungkusAble: CustomMenuCategoryBodyDTO['isBungkusAble'];
-  maxSpicy: CustomMenuCategoryBodyDTO['maxSpicy'];
-};
-
-const customMenuCompositionBodySchema = z.object({
-  customMenuCategoryId: z.string(),
-  name: z.string().nonempty().max(80),
-  description: z.string().nonempty().max(3000),
-  price: z.number().positive(),
-  images: z.array(z.string()).min(1).max(5),
-  stock: z.number().default(0).optional(),
-});
-
-export type CustomMenuCompositionBodyDTO = z.infer<typeof customMenuCompositionBodySchema>;
-export type CustomMenuCompositionResponseDTO = {
-  _id: ICustomMenuComposition['_id'];
-  customMenuCategoryId: CustomMenuCompositionBodyDTO['customMenuCategoryId'];
-  name: CustomMenuCompositionBodyDTO['name'];
-  description: CustomMenuCompositionBodyDTO['description'];
-  price: CustomMenuCompositionBodyDTO['price'];
-  stock: CustomMenuCompositionBodyDTO['stock'];
-  images: CustomMenuCompositionBodyDTO['images'];
-};
-export type GetCustomMenuCompositionsWithPaginated = {
-  customMenuCompositions: Pick<CustomMenuCompositionResponseDTO, '_id' | 'name' | 'price'>[];
-  pages: number;
-  total: number;
-}
+import * as DTO from './types';
 
 const getAllCustomMenuCategory = async (req: Request):
-  Promise<Pick<CustomMenuCategoryResponseDTO, '_id' | 'name'>[] | Error> => {
+  Promise<Pick<DTO.CustomMenuCategoryResponse, '_id' | 'name'>[] | Error> => {
   const { _id: restaurantId } = req.user as { _id: IRestaurant['_id'] };
   try {
     const customMenuCategories = await CustomMenuCategory.find({
       restaurantId
     });
 
-    const result: Pick<CustomMenuCategoryResponseDTO, '_id' | 'name'>[] =
+    const result: Pick<DTO.CustomMenuCategoryResponse, '_id' | 'name'>[] =
       customMenuCategories.map((item) => ({
         _id: item._id,
         name: item.name,
@@ -69,8 +32,8 @@ const getAllCustomMenuCategory = async (req: Request):
 const createCustomMenuCategory = async (req: Request): Promise<ICustomMenuCategory['_id'] | Error> => {
   const { _id: restaurantId } = req.user as { _id: IRestaurant['_id'] };
   try {
-    const body: CustomMenuCategoryBodyDTO =
-      customMenuCategoryBodySchema.parse(req.body);
+    const body: DTO.CustomMenuCategoryBody =
+      DTO.customMenuCategoryBodySchema.parse(req.body);
 
     const { name, isBungkusAble, maxSpicy } = body;
     const customMenuCategory = await CustomMenuCategory.create({
@@ -93,7 +56,7 @@ const createCustomMenuCategory = async (req: Request): Promise<ICustomMenuCatego
   }
 };
 
-const getSpecificCustomMenuCategory = async (req: Request): Promise<CustomMenuCategoryResponseDTO | Error> => {
+const getSpecificCustomMenuCategory = async (req: Request): Promise<DTO.CustomMenuCategoryResponse | Error> => {
   const { _id: restaurantId } = req.user as { _id: IRestaurant['_id'] };
   try {
     const { categoryId } = req.params;
@@ -116,7 +79,7 @@ const getSpecificCustomMenuCategory = async (req: Request): Promise<CustomMenuCa
         ? customMenuCategorySpicyLevelExist.maxSpicy
         : null;
 
-    const result: CustomMenuCategoryResponseDTO = {
+    const result: DTO.CustomMenuCategoryResponse = {
       _id: customMenuCategory._id,
       name: customMenuCategory.name,
       isBungkusAble: customMenuCategory.isBungkusAble,
@@ -133,7 +96,7 @@ const getSpecificCustomMenuCategory = async (req: Request): Promise<CustomMenuCa
 };
 
 const updateCustomMenuCategory = async (req: Request):
-  Promise<CustomMenuCategoryResponseDTO['_id'] | Error> => {
+  Promise<DTO.CustomMenuCategoryResponse['_id'] | Error> => {
   const { _id: restaurantId } = req.user as { _id: IRestaurant['_id'] };
   try {
     const { categoryId } = req.params;
@@ -141,8 +104,8 @@ const updateCustomMenuCategory = async (req: Request):
       throw new BadRequest('Invalid Request. Category id is undefined. Please check your input data.');
     }
 
-    const body: CustomMenuCategoryBodyDTO =
-      customMenuCategoryBodySchema.parse(req.body);
+    const body: DTO.CustomMenuCategoryBody =
+      DTO.customMenuCategoryBodySchema.parse(req.body);
 
     const updatedCustomMenuCategory = await CustomMenuCategory
       .findOneAndUpdate({ _id: categoryId, restaurantId }, {
@@ -155,23 +118,19 @@ const updateCustomMenuCategory = async (req: Request):
 
     const customMenuCategoryHasSpicyLevel =
       await CustomMenuCategorySpicyLevel.findOne({ customMenuCategoryId: categoryId });
-    if (customMenuCategoryHasSpicyLevel) {
-      if (body.maxSpicy) {
-        await CustomMenuCategorySpicyLevel
-          .findOneAndUpdate({ customMenuCategoryId: categoryId }, {
-            maxSpicy: body.maxSpicy,
-          });
-      } else {
-        await CustomMenuCategorySpicyLevel
-          .findOneAndDelete({ customMenuCategoryId: categoryId });
-      }
-    } else {
-      if (body.maxSpicy) {
-        await CustomMenuCategorySpicyLevel.create({
-          customMenuCategoryId: categoryId, 
+    if (customMenuCategoryHasSpicyLevel && body.maxSpicy) {
+      await CustomMenuCategorySpicyLevel
+        .findOneAndUpdate({ customMenuCategoryId: categoryId }, {
           maxSpicy: body.maxSpicy,
         });
-      }
+    } else if (customMenuCategoryHasSpicyLevel && !body.maxSpicy) {
+      await CustomMenuCategorySpicyLevel
+        .findOneAndDelete({ customMenuCategoryId: categoryId });
+    } else if ((!customMenuCategoryHasSpicyLevel && body.maxSpicy)) {
+      await CustomMenuCategorySpicyLevel.create({
+        customMenuCategoryId: categoryId, 
+        maxSpicy: body.maxSpicy,
+      });
     }
 
     return updatedCustomMenuCategory._id;
@@ -184,7 +143,7 @@ const updateCustomMenuCategory = async (req: Request):
   };
 
 const deleteCustomMenuCategory = async (req: Request):
-  Promise<CustomMenuCategoryResponseDTO['_id'] | Error> => {
+  Promise<DTO.CustomMenuCategoryResponse['_id'] | Error> => {
   const { _id: restaurantId } = req.user as { _id: IRestaurant['_id'] };
   try {
     const { categoryId } = req.params;
@@ -196,6 +155,12 @@ const deleteCustomMenuCategory = async (req: Request):
       await CustomMenuComposition.findOne({ customMenuCategoryId: categoryId });
     if (customMenuCategoryHasComponentCustomMenu) {
       throw new BadRequest('Custom Menu Category has at least one composition. Category can not be deleted. Please make sure there is no composition with this category to delete it.');
+    }
+
+    const customMenuCategoryHasSpicyLevel =
+      await CustomMenuCategorySpicyLevel.findOne({ customMenuCategoryId: categoryId });
+    if (customMenuCategoryHasSpicyLevel) {
+      await CustomMenuCategorySpicyLevel.findOneAndDelete({ customMenuCategoryId: categoryId });
     }
 
     const deletedCustomMenuCategory =
@@ -217,7 +182,7 @@ const deleteCustomMenuCategory = async (req: Request):
   };
 
 const getAllCustomMenuComposition = async (req: Request):
-  Promise<GetCustomMenuCompositionsWithPaginated | Error>=> {
+  Promise<DTO.GetCustomMenuCompositionsWithPaginated | Error>=> {
     const { _id: restaurantId } = req.user as { _id: IRestaurant['_id'] };
     const { limit = '10', page = '1', category, name }: {
       limit?: string,
@@ -248,17 +213,21 @@ const getAllCustomMenuComposition = async (req: Request):
           .limit(numberedLimit)
           .skip(numberedLimit * (numberedPage - 1));
       const count = await CustomMenuComposition.countDocuments({ restaurantId, ...filter });
+      const totalPages = Math.ceil(count / numberedLimit);
 
-      const result: GetCustomMenuCompositionsWithPaginated = {
+      if (numberedPage !== 1 && numberedPage > totalPages) {
+        throw new BadRequest('Input page is bigger than total pages. Please check your page query.');
+      }
+
+      const result: DTO.GetCustomMenuCompositionsWithPaginated = {
         customMenuCompositions: customMenuCompositions.map((item) => ({
           _id: item._id,
           name: item.name,
           price: item.price,
         })),
-        pages: Math.ceil(count / numberedLimit),
+        pages: totalPages,
         total: count,
       }
-
       return result;
     } catch (error: any) {
       throw error;
@@ -266,11 +235,11 @@ const getAllCustomMenuComposition = async (req: Request):
   };
 
 const createCustomMenuComposition = async (req: Request):
-  Promise<CustomMenuCompositionResponseDTO['_id'] | Error> => {
-  const { _id: restaurantId } = req.user as { _id: IRestaurant['_id'] }
+  Promise<DTO.CustomMenuCompositionResponse['_id'] | Error> => {
+    const { _id: restaurantId } = req.user as { _id: IRestaurant['_id'] }
     try {
       const body =
-        customMenuCompositionBodySchema.parse(req.body);
+        DTO.customMenuCompositionBodySchema.parse(req.body);
 
       const customMenuCategoryExist =
         await CustomMenuCategory.findOne({ _id: body.customMenuCategoryId, restaurantId });
@@ -278,7 +247,7 @@ const createCustomMenuComposition = async (req: Request):
       if (!customMenuCategoryExist) {
         throw new NotFound('Custom Menu category id is not found. Please input valid category custom menu id.');
       }
-
+      
       const createdCustomMenuComposition =
         await CustomMenuComposition.create({
           restaurantId,
@@ -292,15 +261,18 @@ const createCustomMenuComposition = async (req: Request):
             maxImage: 2,
           })),
         });
-      const { _id: result } = createdCustomMenuComposition;
-      return result;
+        const { _id: result } = createdCustomMenuComposition;
+        return result;
     } catch (error: any) {
+      if (error.name === 'CastError') {
+        throw new NotFound('Custom Menu category id is not found. Please input valid category custom menu id.');
+      }
       throw error;
     }
   };
 
 const getSpecificCustomMenuComposition = async (req: Request):
-  Promise<CustomMenuCompositionResponseDTO | Error> => {
+  Promise<DTO.CustomMenuCompositionResponse | Error> => {
     const { _id: restaurantId } = req.user as { _id: IRestaurant['_id'] };
     try {
       const { compositionId } = req.params;
@@ -311,7 +283,7 @@ const getSpecificCustomMenuComposition = async (req: Request):
         throw new NotFound('Composition custom menu id not found. Please input valid custom menu composition id.');
       }
 
-      const result: CustomMenuCompositionResponseDTO = {
+      const result: DTO.CustomMenuCompositionResponse = {
         _id: customMenuComposition._id,
         name: customMenuComposition.name,
         customMenuCategoryId: customMenuComposition.customMenuCategoryId.toString(),
@@ -334,7 +306,7 @@ const getSpecificCustomMenuComposition = async (req: Request):
   };
 
 const updateCustomMenuComposition = async (req: Request):
-  Promise<CustomMenuCompositionResponseDTO['_id'] | Error> => {
+  Promise<DTO.CustomMenuCompositionResponse['_id'] | Error> => {
     const { _id: restaurantId } = req.user as { _id: IRestaurant['_id'] };
     try {
       const { compositionId } = req.params;
@@ -342,8 +314,8 @@ const updateCustomMenuComposition = async (req: Request):
         throw new BadRequest('Invalid Request. Composistion id is undefined. Please check your input data.');
       }
 
-      const body: CustomMenuCompositionBodyDTO = 
-        customMenuCompositionBodySchema.parse(req.body);
+      const body: DTO.CustomMenuCompositionBody = 
+        DTO.customMenuCompositionBodySchema.parse(req.body);
       
       const customMenuCategoryExist =
         await CustomMenuCategory.findOne({ _id: body.customMenuCategoryId, restaurantId });
@@ -379,7 +351,7 @@ const updateCustomMenuComposition = async (req: Request):
   };
 
 const deleteCustomMenuComposition = async (req: Request):
-  Promise<CustomMenuCompositionResponseDTO['_id'] | Error> => {
+  Promise<DTO.CustomMenuCompositionResponse['_id'] | Error> => {
     const { _id: restaurantId } = req.user as { _id: IRestaurant['_id'] };
     try {
       const { compositionId } = req.params;
